fix: split ListPlans into a container and a component
